[#42] feat (AddRecipeScreen) -> implement ingredient title and portions view

+++ app/src/main/java/de/karzek/diettracker/data/cache/model/RecipeEntity.java
@@ -20,7 +20,7 @@ public class RecipeEntity extends RealmObject {
     private int id;
     private String title;
     private byte[] photo;
-    private int portions;
+    private float portions;
     private RealmList<IngredientEntity> ingredients;
     private RealmList<PreparationStepEntity> steps;
 }
